chore: Move send invite endpoint to endpoints section of app
