+++ webapp/admin/views.py
@@ -1,15 +1,10 @@
 # Packages
 import os
-import json
 import flask
 from flask import make_response
-from canonicalwebteam.exceptions import (
-    StoreApiResponseErrorList,
-)
 from canonicalwebteam.store_api.dashboard import Dashboard
 from canonicalwebteam.store_api.publishergw import PublisherGW
 from canonicalwebteam.store_api.devicegw import DeviceGW
-from flask.json import jsonify
 
 # Local
 from webapp.decorators import login_required, exchange_required
@@ -42,31 +37,6 @@ def get_brand_id(session, store_id):
 @exchange_required
 def get_admin(path):
     return flask.render_template("admin/admin.html", **context)
-
-
-@admin.route("/api/store/<store_id>/invite", methods=["POST"])
-@login_required
-@exchange_required
-def post_invite_members(store_id):
-    members = json.loads(flask.request.form.get("members"))
-
-    res = {}
-
-    try:
-        dashboard.invite_store_members(flask.session, store_id, members)
-        res["msg"] = "Changes saved"
-    except StoreApiResponseErrorList as api_response_error_list:
-        msgs = [
-            f"{error.get('message', 'An error occurred')}"
-            for error in api_response_error_list.errors
-        ]
-
-        msgs = list(dict.fromkeys(msgs))
-
-        for msg in msgs:
-            flask.flash(msg, "negative")
-
-    return jsonify(res)
 
 
 # -------------------- FEATURED SNAPS AUTOMATION ------------------
